# Use random number for pivot