Before:
```
$ go test -bench . -benchmem -count=1                                                                                                                                                     git:master*
goos: darwin
goarch: amd64
pkg: github.com/bluele/psort
BenchmarkSort-4                              300           4020937 ns/op              64 B/op          2 allocs/op
BenchmarkPSort-4                            3000            408808 ns/op              64 B/op          2 allocs/op
BenchmarkPSortLowDisc-4                     1000           1448836 ns/op              64 B/op          2 allocs/op
BenchmarkPSortReverseSorted-4                100          18517292 ns/op              64 B/op          2 allocs/op
PASS
ok      github.com/bluele/psort 11.633s
```

After:
```
$ go test -bench . -benchmem -count=1                                                                                                                                        git:improve-performance*
goos: darwin
goarch: amd64
pkg: github.com/bluele/psort
BenchmarkSort-4                              500           3336768 ns/op              64 B/op          2 allocs/op
BenchmarkPSort-4                            5000            478204 ns/op              64 B/op          2 allocs/op
BenchmarkPSortLowDisc-4                     1000           1660928 ns/op              64 B/op          2 allocs/op
BenchmarkPSortReverseSorted-4               5000            244453 ns/op              64 B/op          2 allocs/op
PASS
ok      github.com/bluele/psort 14.597s
```